TE5-T519_ Navigation Added

+++ IOS_DEMO/IOS_DEMO/TableViewController/TableViewController.swift
@@ -81,7 +81,7 @@ extension TableViewController:UITableViewDataSource {
         return UITableViewCell()
     }
     func tableView(_ tableView: UITableView, heightForHeaderInSection section: Int) -> CGFloat {
-        return Constants.FIFTY
+        return CGFloat(Constants.FIFTY)
     }
     func tableView(_ tableView: UITableView, titleForHeaderInSection section: Int) -> String? {
         return "Section \(section)"
